[CREATE] NoticeAdd mvvm architecture
